WebSockets/tests/submissions/Opera/interfaces/WebSocket/readyState/006.html

<!doctype html>
<title>WebSockets: getting readyState in open</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+':80/ws/echo');
  ws.onopen = t.step_func(function(e) {
    assert_equals(ws.readyState, ws.OPEN);
    ws.onclose = debug;
    ws.close();
    t.done();
  });
  ws.onerror = ws.onmessage = ws.onclose = t.step_func(function() {assert_unreached()});
});
</script>
